Fix error handling documentation in README file
This documentation erroneously states that the gem will raise Faraday
errors, but this hasn't been true for some time. In 2.9.0 it raised
subclasses of Onfido::OnfidoError, and now in 3.x it raises
Onfido::ApiError.

As well as the type being different, the Onfido::ApiError does not
include accessors for the type or fields as specified.